Add files via upload
This is the English version of the Investment Optimizer. It considers investment parameters and cost parameters. In addition to fixed-income investments, volatile returns (such as those found in equity funds or ETFs) can also be simulated. The results are presented both in tabular and graphical form over the entire investment period.